refactor: code split repository entities
