+++ src/common/decorators/api-paginated-response.decorator.ts
@@ -1,7 +1,7 @@
 import { applyDecorators, Type } from "@nestjs/common";
 import { ApiExtraModels, ApiOkResponse, getSchemaPath } from "@nestjs/swagger";
 import { PageDto } from "../dtos/page.dto";
-import { Repo } from "../../repo/repo.entity";
+import { Repo } from "../../repo/entities/repo.entity";
 
 export const ApiPaginatedResponse = <TModel extends Type>(
   model: TModel,
