+++ src/repo/entities/repo.to.user.votes.entity.ts
@@ -7,13 +7,13 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn
 } from "typeorm";
-import { User } from "../user/user.entity";
+import { User } from "../../user/user.entity";
 import { Repo } from "./repo.entity";
 
 @Entity({
-  name: "users_to_repos_stars",
+  name: "users_to_repos_votes",
 })
-export class RepoToUserStars {
+export class RepoToUserVotes {
   @PrimaryGeneratedColumn()
   public id!: number;
 
@@ -40,14 +40,14 @@ export class RepoToUserStars {
   })
   deleted_at: Date;
 
-  @ManyToOne(() => User, (user) => user.repoToUserStars)
+  @ManyToOne(() => User, (user) => user.repoToUserVotes)
   @JoinColumn({
     name: "user_id",
     referencedColumnName: "id",
   })
   public user!: User;
 
-  @ManyToOne(() => Repo, (repo) => repo.repoToUserStars)
+  @ManyToOne(() => Repo, (repo) => repo.repoToUserVotes)
   @JoinColumn({
     name: "repo_id",
     referencedColumnName: "id",
